+++ src/gui/pages/03_Search.py
@@ -1,0 +1,4 @@
+from src.gui.gui_utils import get_log
+
+
+log = get_log()
